Don't use schemaless URLS where not appropriate
Schemaless URLs can only be used in certain places in the browser:

 * http://stackoverflow.com/a/28417193

In the context of requesting a resource using curl, wget, file_get_contents, etc,
there is no way to infer whether the request is supposed to be over https or http.
As such, this change forces the use of http for such requests.